+++ src/api/exercise/InviteGuest.ts
@@ -1,14 +1,12 @@
-import { useQuery } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import api from "../api";
+import type { inviteGuestRequestDTO } from "../../types/InviteGuest";
 
-const axios = api;
-
-//guest가져오기
+//게스트 정보 가져오기
 export const getInviteGuestList = async (exerciseId: number) => {
-  const { data } = await axios.get(`/api/exercises/${exerciseId}/guests`);
+  const { data } = await api.get(`/api/exercises/${exerciseId}/guests`);
   return data;
 };
-//guset 초대하기
 
 export const useInviteGuest = (exerciseId: number) => {
   return useQuery({
@@ -17,5 +15,30 @@ export const useInviteGuest = (exerciseId: number) => {
     queryFn: () => getInviteGuestList(exerciseId),
 
     select: res => res.data,
+  });
+};
+
+//게스트 초대하기
+export const postInviteGuest = async (
+  exerciseId: number,
+  body: inviteGuestRequestDTO,
+) => {
+  return await api.post(`/api/exercises/${exerciseId}/guests`, body);
+};
+
+export const useInviteForm = (exerciseId: number, onSuccess: () => void) => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: (body: inviteGuestRequestDTO) =>
+      postInviteGuest(exerciseId, body),
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["inviteGuest", exerciseId],
+      });
+      onSuccess();
+    },
+    onError: err => {
+      console.log(err);
+    },
   });
 };
