+++ src/types/InviteGuest.ts
@@ -1,0 +1,5 @@
+export type inviteGuestRequestDTO = {
+  guestName: string;
+  gender: string;
+  level: string;
+};
